+++ jira_board_mini/src/components/TaskList.jsx
@@ -1,9 +1,9 @@
-import { useContext, useState } from "react"
+import { useContext } from "react"
 import { StyledTaskList } from "./styles/TaskList.styled"
 import TaskCard from "./TaskCard"
 import { CardContext } from "../context/CardContext"
 import { GlobalContext } from "../context/GlobalContext"
-import useCalculateDraggableStates from "../hooks/useCalculateDraggableStates"
+import { useAssignee } from "../hooks/useAssignee"
 
 function TaskList({main_state, cross_state}) {
   const {
@@ -17,6 +17,9 @@ function TaskList({main_state, cross_state}) {
     Assignees,
     areFiltersActive
   } = useContext(CardContext)
+
+  // const {areFiltersActive} = useAssignee()
+  // console.log("TaskList", areFiltersActive)
 
   const {usable_card_width, usable_card_height} = useContext(GlobalContext)
 
@@ -46,6 +49,7 @@ function TaskList({main_state, cross_state}) {
   let renderCards = Cards.filter(card => card.cross_status === cross_state && card.priority === main_state)
 
   if(areFiltersActive){
+    console.log("Filter Active")
     //go over assignees and see what all are active : filter stage 2
     renderCards = renderCards.filter(card => Assignees[card.assignee].isFilterActive)
   }
