+++ jira_board_mini/src/context/CardContext.jsx
@@ -1,10 +1,13 @@
 import { createContext, useEffect, useState } from "react";
 import getCardsAPI from "../services/getCards"
 import { addCardAPI, deleteCardAPI } from "../services/setCards"
+import { useAssignee } from "../hooks/useAssignee";
 
 export const CardContext = createContext();
 
 export const CardProvider = ({children, initialAssignees = {}}) => {
+
+    const {Assignees, addAssignee, removeAssignee, toggleAssigneeFilter, areFiltersActive} = useAssignee()
 
     useEffect(() => {
         const fetchCards = async () => {
@@ -13,78 +16,30 @@ export const CardProvider = ({children, initialAssignees = {}}) => {
 
             if(isCached)
                 return
-            setAssignees(prevAssignees => {
-                const newAssignees = {...prevAssignees};
-                
-                fetchedCards.forEach(card => {
-                    if (card.assignee in newAssignees) {
-                        newAssignees[card.assignee] = {
-                            count: newAssignees[card.assignee].count + 1,
-                            isFilterActive: newAssignees[card.assignee].isFilterActive
-                        };
-                    } else {
-                        newAssignees[card.assignee] = {
-                            count: 1,
-                            isFilterActive: false
-                        };
-                    }
-                });
-                return newAssignees;
-            });
+
+            fetchedCards.forEach(card => {
+                addAssignee(card.assignee)
+            })
         }
         fetchCards()
     }, [])
 
     const [Cards, setCards] = useState([])
-    const [Assignees, setAssignees] = useState(() => {
 
-        if(initialAssignees.length > 0){
-            return initialAssignees
-        }
-
-        const savedAssignees = localStorage.getItem('assignees')
-        return savedAssignees ? JSON.parse(savedAssignees) : {}
-    });
     const [DraggedCard, setDraggedCard] = useState(null)
     const [IsDragActive, setIsDragActive] = useState(false)
     const [DraggableStates, setDraggableStates] = useState({})
 
-    const [areFiltersActive, setAreFiltersActive] = useState(false)
-
     useEffect(() => {
         localStorage.setItem('cards', JSON.stringify(Cards))
     }, [Cards])
-    
-    useEffect(() => {
-        localStorage.setItem('assignees', JSON.stringify(Assignees))
-    }, [Assignees])
 
     const addCard = async (card, _priority, _cross_status) => {
         const newCard = {...card, priority: Number(_priority), cross_status: Number(_cross_status)}
 
         setCards(prevCards => [...prevCards, newCard])
-        
-        setAssignees(prevAssignees => {
-            if(newCard.assignee in prevAssignees){
-                return {
-                    ...prevAssignees,
-                    [newCard.assignee] : {
-                        count: prevAssignees[newCard.assignee].count + 1,
-                        isFilterActive: prevAssignees[newCard.assignee].isFilterActive
-                    }
-                 }
-            }
-                
-            else{
-                return {
-                    ...prevAssignees,
-                    [newCard.assignee] : {
-                        count: 1,
-                        isFilterActive: false
-                    }
-                }
-            }
-        })
+
+        addAssignee(newCard.assignee)
 
         const response = await addCardAPI(newCard)
         console.log(response)
@@ -93,58 +48,12 @@ export const CardProvider = ({children, initialAssignees = {}}) => {
         const newCards = Cards.filter(card => card.id !== id) //O(n)
         const delAssignee = Cards.find(card => card.id === id).assignee //O(n) ; earlier using CardsIndex[id].assignee, it was O(1)
 
-        setAssignees(prevAssignees => {
-            if (!prevAssignees[delAssignee]) return prevAssignees
-
-            if(prevAssignees[delAssignee].count > 1){
-                return {
-                    ...prevAssignees,
-                    [delAssignee]: {
-                        ...prevAssignees[delAssignee],
-                        count: prevAssignees[delAssignee].count - 1,
-                        isFilterActive: prevAssignees[delAssignee].isFilterActive
-                    }
-                }
-            }
-            else{
-                const {[delAssignee] : _, ...rest} = prevAssignees
-                return rest
-            }
-        })
+        removeAssignee(delAssignee)
 
         setCards(newCards)
 
         const response = await deleteCardAPI(id)
         console.log(response)
-    }
-
-    const toggleAssigneeFilter = (toggleAssignee) => {
-
-        if(Assignees[toggleAssignee].isFilterActive === false){
-            setAreFiltersActive(true)
-        }
-        else{
-            //basically u are toggling toggleAssignee off; check if all other filters are off as well
-            let allFiltersOff = true
-            Object.keys(Assignees).forEach(assigneeKey => {
-                if(assigneeKey !== toggleAssignee && Assignees[assigneeKey].isFilterActive === true){
-                    allFiltersOff = false
-                    return
-                }
-            })
-            if(allFiltersOff){
-                setAreFiltersActive(false)
-            }
-        }
-        setAssignees(prevAssignees => {
-            const newAssignees = {...prevAssignees} //O(n)
-            newAssignees[toggleAssignee] = {
-                ...newAssignees[toggleAssignee],
-                isFilterActive: !newAssignees[toggleAssignee].isFilterActive
-            }
-
-            return newAssignees
-        })
     }
 
     return (
