+++ jira_board_mini/src/hooks/useAssignee.js
@@ -1,0 +1,96 @@
+import { use, useEffect } from "react";
+import { useState } from "react";
+import { set } from "react-hook-form";
+
+export const useAssignee = (initialAssignees = {}) => {
+    const [Assignees, setAssignees] = useState(() => {
+
+        if(initialAssignees.length > 0){
+            return initialAssignees
+        }
+        const savedAssignees = localStorage.getItem('assignees')
+        return savedAssignees ? JSON.parse(savedAssignees) : {}
+    });
+
+    const [areFiltersActive, setAreFiltersActive] = useState(false)
+
+    useEffect(() => {
+        localStorage.setItem('assignees', JSON.stringify(Assignees))
+    }, [Assignees])
+
+    const addAssignee = (assignee) => {
+        setAssignees(prevAssignees => {
+            if (assignee in prevAssignees) {
+                return {
+                    ...prevAssignees,
+                    [assignee]: {
+                        count: prevAssignees[assignee].count + 1,
+                        isFilterActive: prevAssignees[assignee].isFilterActive
+                    }
+                }
+            }
+            else{
+                return {
+                    ...prevAssignees,
+                    [assignee]: {
+                        count: 1,
+                        isFilterActive: false
+                    }
+                }
+            }
+        })
+    }
+
+    const removeAssignee = (assignee) => {
+        setAssignees(prevAssignees => {
+            if(!prevAssignees[assignee]){
+                return prevAssignees
+            }
+
+            if (prevAssignees[assignee].count === 1) {
+                const newAssignees = { ...prevAssignees }
+                delete newAssignees[assignee]
+                return newAssignees
+            }
+            return {
+                ...prevAssignees,
+                [assignee]: {
+                    count: prevAssignees[assignee].count - 1,
+                    isFilterActive: prevAssignees[assignee].isFilterActive
+                }
+            }
+        })
+    }
+
+    const toggleAssigneeFilter = (toggleAssignee) => {
+
+        if(Assignees[toggleAssignee].isFilterActive === false){
+            setAreFiltersActive(true)
+        }
+        else{
+            //basically u are toggling toggleAssignee off; check if all other filters are off as well
+            let allFiltersOff = true
+            Object.keys(Assignees).forEach(assigneeKey => {
+                if(assigneeKey !== toggleAssignee && Assignees[assigneeKey].isFilterActive === true){
+                    allFiltersOff = false
+                    return
+                }
+            })
+            if(allFiltersOff){
+                setAreFiltersActive(false)
+            }
+        }
+        console.log(areFiltersActive)
+        setAssignees(prevAssignees => {
+            const newAssignees = {...prevAssignees} //O(n)
+            newAssignees[toggleAssignee] = {
+                ...newAssignees[toggleAssignee],
+                isFilterActive: !newAssignees[toggleAssignee].isFilterActive
+            }
+
+            return newAssignees
+        })
+    }
+
+    return {Assignees, setAssignees, addAssignee, removeAssignee, toggleAssigneeFilter, areFiltersActive}
+}
