+++ jira_board_mini/src/hooks/useCalculateDraggableStates.js
@@ -8,7 +8,7 @@ const useCalculateDraggableStates = ({main_state, cross_state}) => {
     const {card_move_rules_horizontal, card_move_rules_vertical} = useContext(GlobalContext)
     const {setDraggableStates} = useContext(CardContext)
 
-    console.log("main_state", main_state, "cross_state", cross_state)
+    // console.log("main_state", main_state, "cross_state", cross_state)
 
     const DraggableStates = useMemo(() => {
         const newDraggableStates = {}
@@ -33,7 +33,7 @@ const useCalculateDraggableStates = ({main_state, cross_state}) => {
             newDraggableStates[i][cross_state] = true
         }
 
-        console.log("Calculating Draggable States:", main_state, cross_state, newDraggableStates)
+        // console.log("Calculating Draggable States:", main_state, cross_state, newDraggableStates)
         return newDraggableStates
     }, [main_state, cross_state])
